Move set functions into set.js.
