+++ packages/stochator/index.js
@@ -1,6 +1,12 @@
 import seedrandom from "seedrandom";
 import {randomBoundedFloat} from "./float";
 import {randomBoundedInteger} from "./integer";
+import {
+    randomSetMember,
+    randomSetMemberWithoutReplacement,
+    randomWeightedSetMember,
+    shuffleSet
+} from "./set";
 
 const isType = (type) => {
     return (arg) => Object.prototype.toString.call(arg) == `[object ${ type }]`
@@ -36,36 +42,6 @@ const randomCharacter = (prng, lowercase) => {
     const [min, max] = lowercase ? [97, 122] : [65, 90];
     const mutator = (charCode) => String.fromCharCode(charCode);
     return new Stochator({ kind: "integer", min, max, prng }, mutator).next;
-};
-
-const randomSetMember = (prng, values) => {
-    const max = values.length - 1;
-    return values[randomBoundedInteger(prng, 0, max)];
-};
-
-const randomSetMemberWithoutReplacement = (prng, values) => {
-    if (values.length > 0) {
-        const index = randomBoundedInteger(prng, 0, values.length - 1);
-        const value = values[index];
-        values.splice(index, 1);
-        return value;
-    }
-};
-
-const randomWeightedSetMember = (prng, values, weights) => {
-    let [member, weightSum, float] = [undefined, 0, randomBoundedFloat(prng)];
-    values.forEach((value, index) => {
-        if (member) {
-            return;
-        }
-        const weight = weights[index];
-        if (float <= weightSum + weight && float >= weightSum) {
-            member = value;
-        }
-        weightSum += weight;
-    });
-
-    return member;
 };
 
 const inverseNormalCumulativeDistribution = (probability) => {
@@ -113,18 +89,6 @@ const inverseNormalCumulativeDistribution = (probability) => {
     const denominator = sum(denomCoeffcients.map(mapMaxExp(denomMaxExponent))) + 1;
 
     return coefficient * numerator / denominator;
-};
-
-const shuffleSet = (prng, values) => {
-    let valuesRef = [...values];
-    for (index of range(0, valuesRef.length)) {
-        randomIndex = randomBoundedInteger(prng, 0, index);
-
-        tmp = valuesRef[index];
-        valuesRef[index] = valuesRef[randomIndex];
-        valuesRef[randomIndex] = tmp;
-    }
-    return valuesRef;
 };
 
 const floatGenerator = (prng, min, max, mean, stdev) => {
